Fix bug: SEQ writes the wrong length of injection.
Bug was introduced in commit 9655a87354d089de716bc417cada90c02fb078b1.

Fixed by keeping track of "number of bits written to sequence" and
updating that value in the specific case that the first command is an
injection (dealt with in a special way).

Along the way, changed the _block_length member variables to a dict,
_block_lengths[...]. Acceptable keys are currently 'pixel', 'global',
and 'inject'.